this is gonna take forever to properly code...
