+++ Content.Shared/_Goobstation/SpaceWhale/SharedTailedEntitySystem.cs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+using System.Linq;
 using System.Numerics;
 using Content.Shared.Interaction;
 using Content.Shared.Interaction.Events;
@@ -7,6 +8,8 @@ using Content.Shared._Goobstation.SpaceWhale.Weapons;
 using Content.Shared.Weapons.Melee;
 using Robust.Shared.Physics.Events;
 using Robust.Shared.Map;
+using Robust.Shared.Physics.Systems;
+using Robust.Shared.Timing;
 
 namespace Content.Shared._Goobstation.SpaceWhale;
 
@@ -14,39 +17,84 @@ public abstract class SharedTailedEntitySystem : EntitySystem
 {
     [Dependency] protected readonly SharedTransformSystem TransformSystem = default!;
 
+    [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly SharedInteractionSystem _interaction = default!;
+    [Dependency] private readonly EntityLookupSystem _look = default!;
+
+    private EntityQuery<TailedEntitySegmentComponent> _segmentQuery;
+    private readonly HashSet<Entity<TailedEntitySegmentComponent>> _lookSegments = new();
 
     public override void Initialize()
     {
         base.Initialize();
 
+        _segmentQuery = GetEntityQuery<TailedEntitySegmentComponent>();
         UpdatesOutsidePrediction = true;
 
         SubscribeLocalEvent<TailedEntityComponent, GetLightAttackRangeEvent>(OnGetRange);
         SubscribeLocalEvent<TailedEntityComponent, MeleeInRangeEvent>(OnInRange);
         SubscribeLocalEvent<TailedEntityComponent, AttackAttemptEvent>(OnAttackAttempt);
+        SubscribeLocalEvent<TailedEntityComponent, PreventCollideEvent>(OnPreventCollide);
+        SubscribeLocalEvent<TailedEntityComponent, TailedEntityForceContractEvent>(OnForceContract);
+    }
+
+    private void OnForceContract(Entity<TailedEntityComponent> ent, ref TailedEntityForceContractEvent args)
+    {
+        args.Handled = true;
+        ent.Comp.PreventSegmentCollide = true;
+
+        var pos = TransformSystem.GetMapCoordinates(ent);
+
+        foreach (var data in ent.Comp.TailSegments)
+        {
+            if (!TryGetEntity(data.Segment, out var segment) || !_segmentQuery.TryComp(segment.Value, out var comp))
+                continue;
+
+            comp.Coords = pos;
+            TransformSystem.SetMapCoordinates(segment.Value, pos);
+            Dirty(segment.Value, comp);
+
+            data.WorldPosition = pos.Position;
+        }
+
+        Dirty(ent);
+    }
+
+    private void OnPreventCollide(Entity<TailedEntityComponent> ent, ref PreventCollideEvent args)
+    {
+        var other = args.OtherEntity;
+
+        var index = ent.Comp.TailSegments.FindIndex(x => TryGetEntity(x.Segment, out var e) && e == other);
+
+        if (index < 0)
+            return;
+
+        if (ent.Comp.PreventSegmentCollide ||
+            ent.Comp.PreventFirstSegmentsCollideAmount < 1 && index < ent.Comp.PreventFirstSegmentsCollideAmount)
+            args.Cancelled = true;
     }
 
     private void OnAttackAttempt(Entity<TailedEntityComponent> ent, ref AttackAttemptEvent args)
     {
-        if (args.Target is { } target && ent.Comp.TailSegments.Contains(target))
+        if (args.Target is { } target &&
+            ent.Comp.TailSegments.Any(x => TryGetEntity(x.Segment, out var e) && e == target))
             args.Cancel();
     }
 
     private void OnInRange(Entity<TailedEntityComponent> ent, ref MeleeInRangeEvent args)
     {
-        if (args.Handled)
+        if (args.Handled || !ent.Comp.MeleeAttackWithSegments)
             return;
 
         args.Handled = true;
 
-        var segments = new List<EntityUid> { ent };
-        segments.AddRange(ent.Comp.TailSegments);
+        var segments = ent.Comp.TailSegments.Select(x => GetEntity(x.Segment)).ToList();
+        segments.Insert(0, ent);
 
         foreach (var segment in segments)
         {
             if (!Exists(segment))
-                return;
+                continue;
 
             if (!CheckInRange(segment, segments, ref args))
                 continue;
@@ -72,51 +120,81 @@ public abstract class SharedTailedEntitySystem : EntitySystem
 
     private void OnGetRange(Entity<TailedEntityComponent> ent, ref GetLightAttackRangeEvent args)
     {
-        if (!TryComp(ent, out MeleeWeaponComponent? melee))
+        if (!ent.Comp.MeleeAttackWithSegments || !TryComp(ent, out MeleeWeaponComponent? melee))
             return;
 
         args.Cancel = true;
         args.Range = ent.Comp.TailSegments.Count + melee.Range;
     }
 
+    public override void FrameUpdate(float frameTime)
+    {
+        base.FrameUpdate(frameTime);
+
+        if (_timing.ApplyingState)
+            return;
+
+        var query = EntityQueryEnumerator<TailedEntitySegmentComponent>();
+        while (query.MoveNext(out var uid, out var segment))
+        {
+            ResetSegmentPosition((uid, segment));
+        }
+    }
+
     public override void Update(float frameTime)
     {
         base.Update(frameTime);
 
+        if (_timing.ApplyingState)
+            return;
+
         var query = EntityQueryEnumerator<TailedEntityComponent, TransformComponent>();
         while (query.MoveNext(out var uid, out var comp, out var xform))
         {
+            if (comp.TailSegments.Count == 0)
+                continue;
+
             UpdateTailPositions((uid, comp, xform));
-            UpdateTailLayers((uid, comp));
+            UpdateCollision((uid, comp, xform));
         }
     }
 
-    protected virtual void UpdateTailLayers(Entity<TailedEntityComponent> ent) { }
-
-    private void UpdateTailPositions(Entity<TailedEntityComponent, TransformComponent> ent)
+    private void UpdateCollision(Entity<TailedEntityComponent, TransformComponent> ent)
     {
+        if (!ent.Comp1.ShouldCollideWithSegments || !ent.Comp1.PreventSegmentCollide)
+            return;
+
+        _lookSegments.Clear();
+        _look.GetEntitiesInRange(ent.Comp2.Coordinates, ent.Comp1.HeadRadius, _lookSegments, LookupFlags.Dynamic);
+        if (ent.Comp1.TailSegments.Any(x =>
+                TryGetEntity(x.Segment, out var e) && _lookSegments.Any(y => e == y.Owner)))
+            return;
+
+        ent.Comp1.PreventSegmentCollide = false;
+        Dirty(ent, ent.Comp1);
+    }
+
+    protected void UpdateTailPositions(Entity<TailedEntityComponent, TransformComponent> ent)
+    {
+        if (_timing.ApplyingState)
+            return;
+
         var (uid, comp, xform) = ent;
 
-        // Use the head's world position to determine whether anything moved.
         var headPos = TransformSystem.GetWorldPosition(xform);
-        if (headPos == comp.LastPos)
+
+        if (headPos == ent.Comp1.LastPos)
             return;
+
+        ent.Comp1.LastPos = headPos;
 
         Angle? headRot = null;
         for (var i = 0; i < comp.TailSegments.Count; i++)
         {
-            var segment = comp.TailSegments[i];
+            var data = comp.TailSegments[i];
 
-            if (TerminatingOrDeleted(segment))
-                continue;
-
-            EntityUid? next = i <= 0 ? uid : comp.TailSegments[i - 1];
-
-            if (TerminatingOrDeleted(next))
-                continue;
-
-            var segPos = TransformSystem.GetWorldPosition(segment);
-            var nextPos = TransformSystem.GetWorldPosition(next.Value);
+            var segPos = data.WorldPosition;
+            var nextPos = i <= 0 ? headPos : comp.TailSegments[i - 1].WorldPosition;
             var nextRot = Angle.FromWorldVec(nextPos - segPos);
             headRot ??= nextRot;
 
@@ -137,12 +215,30 @@ public abstract class SharedTailedEntitySystem : EntitySystem
                 desiredPos = nextPos - nextRot.ToWorldVec() * comp.Spacing;
             }
 
-            TransformSystem.SetMapCoordinates(segment, new MapCoordinates(desiredPos, xform.MapID));
-            TransformSystem.SetWorldRotation(segment, nextRot);
+            comp.TailSegments[i].WorldPosition = desiredPos;
+
+            if (!TryGetEntity(data.Segment, out var segment) ||
+                !_segmentQuery.TryComp(segment.Value, out var segmentComp))
+                continue;
+
+            segmentComp.Coords = new MapCoordinates(desiredPos, xform.MapID);
+            segmentComp.WorldRotation = nextRot;
+
+            Dirty(segment.Value, segmentComp);
+
+            ResetSegmentPosition((segment.Value, segmentComp));
         }
 
-        if (headRot is { } rot)
-            TransformSystem.SetWorldRotation(ent.Owner, rot);
-        comp.LastPos = headPos;
+        if (comp.HeadFollowSegmentRotation && headRot is { } rot)
+            TransformSystem.SetWorldRotation(uid, rot);
+
+        Dirty(uid, comp);
+    }
+
+    protected void ResetSegmentPosition(Entity<TailedEntitySegmentComponent> segment)
+    {
+        if (segment.Comp.Coords is { } coords)
+            TransformSystem.SetMapCoordinates(segment, coords);
+        TransformSystem.SetWorldRotation(segment, segment.Comp.WorldRotation);
     }
 }
