--- Content.Trauma.Server/PvsOverride/ForcePvsOverrideComponent.cs
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-namespace Content.Trauma.Server.PvsOverride;
-
-/// <summary>
-/// Added to entities to make them pvs override
-/// </summary>
-[RegisterComponent]
-public sealed partial class ForcePvsOverrideComponent : Component;
